requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appdirs=1.4.4=pypi_0
appnope=0.1.0=py36hf537a9a_0
apscheduler=3.6.3=pypi_0
backcall=0.2.0=py_0
backports-zoneinfo=0.2.1=pypi_0
blas=1.0=mkl
ca-certificates=2020.10.14=0
cachetools=4.2.2=pypi_0
certifi=2019.9.11=pypi_0
chardet=3.0.4=pypi_0
colorama=0.4.4=pypi_0
decorator=4.4.2=py_0
idna=2.8=pypi_0
importlib-resources=5.2.2=pypi_0
intel-openmp=2021.3.0=hecd8cb5_3375
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=py36h241746c_0
jedi=0.18.0=py36hecd8cb5_1
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py36_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libsodium=1.0.18=h1de35cc_0
lxml=4.4.1=pypi_0
mkl=2019.4=233
mkl-service=2.3.0=py36h9ed2024_0
mkl_fft=1.3.0=py36ha059aab_0
mkl_random=1.0.4=py36h27c97d8_0
ncurses=6.2=h0a44026_1
numpy=1.17.2=pypi_0
openssl=1.1.1l=h9ed2024_0
pandas=1.0.3=py36h6c726b0_0
parso=0.8.0=py_0
pexpect=4.8.0=py36_0
pickleshare=0.7.5=py36_0
pip=21.0.1=py36hecd8cb5_0
prompt-toolkit=3.0.8=py_0
ptyprocess=0.6.0=py36_0
pygments=2.7.1=py_0
python=3.6.13=h88f2d9e_0
python-dateutil=2.8.0=pypi_0
python-telegram-bot=13.7=pypi_0
pytz=2019.2=pypi_0
pyzmq=19.0.2=py36hb1e8313_1
readline=8.1=h9ed2024_0
requests=2.22.0=pypi_0
rightmove-webscraper=1.0.2=pypi_0
setuptools=52.0.0=py36hecd8cb5_0
six=1.12.0=pypi_0
sqlite=3.36.0=hce871da_0
telegram-send=0.25=pypi_0
tk=8.6.10=hb0a8c7a_0
tornado=6.1=pypi_0
traitlets=4.3.3=py36_0
tzdata=2021a=h5d7bf9c_0
tzlocal=3.0=pypi_0
urllib3=1.25.6=pypi_0
wcwidth=0.2.5=py_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h1de35cc_0
zeromq=4.3.3=hb1e8313_3
zipp=3.5.0=pypi_0
zlib=1.2.11=h1de35cc_3
